feat: ProductPage -> HeaderPage Rename
