--- src/Pages/ProductPage/ProductPage.tsx
@@ -1,14 +1,0 @@
-import { Outlet } from 'react-router-dom';
-
-import ProductPageStyle from './ProductPage.style';
-
-import { Header } from '@Components/.';
-
-export const ProductPage = () => (
-  <>
-    <Header />
-    <ProductPageStyle.Container>
-      <Outlet />
-    </ProductPageStyle.Container>
-  </>
-);
