--- src/Pages/ProductPage/index.ts
@@ -1,1 +1,0 @@
-export * from './ProductPage';
