+++ src/Pages/index.ts
@@ -6,4 +6,5 @@ export * from './LoginPage';
 export * from './MyPage';
 export * from './RegisterPage';
 export * from './FooterPage';
-export * from './ProductPage';
+export * from './HeaderPage';
+export * from './ProductRegisterPage';
